Add ONNX export functionality and extensive converter implementations
- Introduced the `marquetry.onnx_export` module with support for ONNX IR export via `export_onnx`.
- Implemented converters for various operations (e.g., math, array manipulation, activations, normalization, pooling, and regularization).
- Added end-to-end tests for ONNX export, validating graph correctness and runtime inference.

+++ marquetry/layers/activations/dynamic_swish.py
@@ -40,10 +40,10 @@ class DynamicSwish(Layer):
 
     """
 
-    def __init__(self, init_beta=1.):
+    def __init__(self, init_beta=1., dtype=np.float32):
         super().__init__()
 
-        self.beta = Parameter(np.array(init_beta))
+        self.beta = Parameter(np.array(init_beta, dtype=dtype))
         self.init_check = False
 
     def forward(self, x):
